pyproject.toml
[project]
name = "gold-wirewatch"
version = "0.1.0"
description = "Low-latency Gold Futures Asia Session News Wire + Triage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer>=0.12",
  "httpx>=0.27",
  "feedparser>=6.0.11",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "python-dotenv>=1.0",
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "tzdata>=2025.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "mypy>=1.11",
  "ruff>=0.6"
]

[project.scripts]
wirewatch = "gold_wirewatch.cli:app"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
addopts = "-q --cov=gold_wirewatch --cov-report=term-missing --cov-fail-under=90"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
